+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_responses.hpp
@@ -249,7 +249,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake { nam
 
   struct DownloadDataLakeFileResult
   {
-    std::unique_ptr<Azure::IO::BodyStream> Body;
+    std::unique_ptr<Azure::Core::IO::BodyStream> Body;
     int64_t FileSize = int64_t();
     Azure::Core::Http::HttpRange ContentRange;
     Azure::Core::Nullable<Storage::ContentHash> TransactionalContentHash;